+++ src/stats.rs
@@ -1,8 +1,12 @@
-use std::sync::atomic::{AtomicU64, Ordering::Relaxed};
+use std::{
+    sync::atomic::{AtomicU64, Ordering::Relaxed},
+    time::Instant,
+};
 
 /// Contains basic statistics related to a node or a connection.
-#[derive(Default)]
 pub struct Stats {
+    /// The creation time.
+    created: Instant,
     /// The number of all messages sent.
     msgs_sent: AtomicU64,
     /// The number of all messages received.
@@ -11,6 +15,18 @@ pub struct Stats {
     bytes_sent: AtomicU64,
     /// The number of all bytes received.
     bytes_received: AtomicU64,
+}
+
+impl Default for Stats {
+    fn default() -> Self {
+        Self {
+            created: Instant::now(),
+            msgs_sent: Default::default(),
+            msgs_received: Default::default(),
+            bytes_sent: Default::default(),
+            bytes_received: Default::default(),
+        }
+    }
 }
 
 impl Stats {
@@ -24,6 +40,11 @@ impl Stats {
     pub fn register_received_message(&self, size: usize) {
         self.msgs_received.fetch_add(1, Relaxed);
         self.bytes_received.fetch_add(size as u64, Relaxed);
+    }
+
+    /// Returns the creation time.
+    pub fn created(&self) -> Instant {
+        self.created
     }
 
     /// Returns the number of sent messages and their collective size in bytes.